requirements.txt
# package location
--find-links https://download.pytorch.org/whl/torch_stable.html

# packages
blobfile==2.0.2
diffusers==0.23.1
huggingface-hub==0.25.2
fvcore==0.1.5.post20221221
ipython==8.12.2
lmdb==1.4.1
matplotlib==3.7.2
numpy==1.24.4
omegaconf==2.1.1
opencv_python==4.8.1.78
Pillow==10.0.0
Requests==2.31.0
scikit_learn==1.3.0
scipy==1.9.1
setuptools==68.0.0
tqdm==4.66.1
transformers==4.25.1
torch==2.1.0+cu121
torchvision==0.16.0+cu121
pymeshlab==2023.12.post2
tensorboard
lightning
einops
lpips
scikit-image
plyfile
pymeshlab
accelerate